+++ ChangeLog.md
@@ -33,6 +33,7 @@ Bug fixes:
 * Support for spaces in Haddock interface files [fpco/minghc#85](https://github.com/fpco/minghc/issues/85)
 * Temporarily building against a "shadowing" local package? [#992](https://github.com/commercialhaskell/stack/issues/992)
 * Fix Setup.exe name for --upgrade-cabal on Windows [#1002](https://github.com/commercialhaskell/stack/issues/1002)
+* Unlisted dependencies no longer trigger extraneous second build [#838](https://github.com/commercialhaskell/stack/issues/838)
 
 ## 0.1.4.1